refactor(web): app.js から project-shell surface を切り出し最終 5,483 行にする
SPEC-3064 Phase 3 (wave 2)。project-shell-surface.js (1,268) を新設し、
window list / project tabs / picker / open・clone・migration modal glue /
chrome listeners を移管。app.js 6,434 → 5,483 行（開始時 16,120 から
66% 削減）。再代入される appState / viewport は getter 渡しで本文不変、
receive() は薄い委譲（window_list / clone 6 case / migration 4 case）。

検証: run-frontend-unit-tests 797/797、check-frontend-bundle PASS、
cargo test -p gwt --bin gwt embedded_web 104/104。
