+++ scripts/check-frontend-bundle.sh
@@ -60,3 +60,4 @@ node_check crates/gwt/web/branches-cleanup-surface.js
 node_check crates/gwt/web/board-logs-surface.js
 node_check crates/gwt/web/knowledge-kanban-surface.js
 node_check crates/gwt/web/profile-window-surface.js
+node_check crates/gwt/web/project-shell-surface.js
